refactor: upload-art page
